requirements.txt
astroid==4.0.2
autopep8==2.3.2
colorama==0.4.6
colorzero==2.0
dill==0.4.0
gpiozero==2.0.1
isort==7.0.0
lazy-object-proxy==1.12.0
mccabe==0.7.0
numpy==2.3.5
orderly-set==5.5.0
platformdirs==4.5.0
pycodestyle==2.14.0
pygame-ce==2.5.6
pylint==4.0.3
pynput==1.8.1
setuptools==80.9.0
six==1.17.0
toml==0.10.2
tomlkit==0.13.3
wheel==0.45.1
wrapt==2.0.1
